requirements.txt
prometheus_client
PyYAML
Flask
